pause feature

+++ main.cpp
@@ -1,4 +1,3 @@
-// CS335 fall 2015
 // ====================
 // Final Project - Group 2
 // Game: Doodle Jump
@@ -21,6 +20,8 @@ int window_height = 600;
 
 // check key inputs
 int keys[65536];
+
+bool pausegame = false;
 bool killmovement = false;
 
 // Gordon's timer & x11/opengl code
@@ -65,25 +66,25 @@ int main()
 	initXWindows();
 	init_opengl();
 	loadTextures(); 
-	
+
 	cout << "start game" << endl;
-	
+
 	Game game;
-	
+
 	game.setGravity(GRAVITY);
-	
+
 	// set players position
 	game.setPos(window_width/2, window_height/2);
-	
+
 	game.setResolution(window_width, window_height);
-	
-	
+
+
 	srand(time(NULL));
-	
+
 	clock_gettime(CLOCK_REALTIME, &timePause);
 	clock_gettime(CLOCK_REALTIME, &timeStart);
 	clock_gettime(CLOCK_REALTIME, &start);
-	
+
 	while(game.run)
 	{
 		// check input
@@ -97,37 +98,39 @@ int main()
 			// if window resets, then the game should handle this event
 			game.setResolution(window_width, window_height);
 		}
+		//if(!pausegame)
+		//{
+			// asteroids timer
+			// get lastest time
+			clock_gettime(CLOCK_REALTIME, &timeCurrent);
+			timeSpan = timeDiff(&timeStart, &timeCurrent);
+			// start time is now current time;
+			timeCopy(&timeStart, &timeCurrent);
+
+			physicsCountdown += timeSpan;
+		//}	// check for collisions, move player
+			while(physicsCountdown >= physicsRate) {
+				physics(&game);
+
+				physicsCountdown -= physicsRate;
+			}
+
+			// used for sprite timing
+			if(frames > 2)
+			{
+				clock_gettime(CLOCK_REALTIME, &start);
+				frames = 0;
+			}
+			frames++;
+
+			render(&game);
+			glXSwapBuffers(dpy, win);
 		
-		// asteroids timer
-		// get lastest time
-		clock_gettime(CLOCK_REALTIME, &timeCurrent);
-		timeSpan = timeDiff(&timeStart, &timeCurrent);
-		// start time is now current time;
-		timeCopy(&timeStart, &timeCurrent);
-		
-		physicsCountdown += timeSpan;
-		// check for collisions, move player
-		while(physicsCountdown >= physicsRate) {
-			physics(&game);
-				
-			physicsCountdown -= physicsRate;
-		}
-		
-		// used for sprite timing
-		if(frames > 2)
-		{
-			clock_gettime(CLOCK_REALTIME, &start);
-			frames = 0;
-		}
-		frames++;
-			
-		render(&game);
-		glXSwapBuffers(dpy, win);
 	}	
-	
+
 	cleanupXWindows();
 	cleanup_fonts();
-	
+
 	cout << "end game" << endl;
 	return 0;
 }
@@ -141,39 +144,52 @@ int main()
 int check_keys(XEvent *e, Game * game)
 {
 	int key = XLookupKeysym(&e->xkey, 0);
-		
-	
-	killmovement = true;
+
+
+	//killmovement = true;
 	if(e->type == KeyRelease) 
 	{
 		keys[key] = 0;
+		killmovement = true;
 	}
-	
+
 	if(e->type == KeyPress)
 	{
 		keys[key] = 1;
-		
-		if(key != XK_Left || key != XK_Right)
-			killmovement = false;
-		if(key == XK_Escape)
+
+		if(key == XK_p)
 		{
-			game->run = false;
-		}
-		
-		if(key == XK_b)
-		{
-			if(bubbler)
-				bubbler = false;
+			if(pausegame)
+			{
+				pausegame = false;
+			}
 			else
-				bubbler = true;
-		}	
-		
-		if(key == XK_w)
+				pausegame = true;
+		}
+		if(!pausegame)
 		{
-			if(setbackground)
-				setbackground = false;
-			else 
-				setbackground = true;
+			if(key != XK_Left || key != XK_Right || key != XK_space)
+				killmovement = false;
+			if(key == XK_Escape)
+			{
+				game->run = false;
+			}
+
+			if(key == XK_b)
+			{
+				if(bubbler)
+					bubbler = false;
+				else
+					bubbler = true;
+			}	
+
+			if(key == XK_w)
+			{
+				if(setbackground)
+					setbackground = false;
+				else 
+					setbackground = true;
+			}
 		}
 	}	
 	return 0;
@@ -190,7 +206,7 @@ void makeParticle(int x, int y) {
 
 	p->s.center.x = x;
 	p->s.center.y = y;
-	
+
 	p->velocity.x =  1.0 + rnd() * 0.1;
 
 	p->velocity.y = rnd() * 1.0 - 0.5;
@@ -213,7 +229,7 @@ void check_mouse(XEvent *e, Game *game)
 	if (e->type == ButtonPress) {
 		if (e->xbutton.button==1) {
 			//Left button was pressed
-	
+
 			return;
 		}
 		if (e->xbutton.button==3) {
@@ -235,54 +251,60 @@ void check_mouse(XEvent *e, Game *game)
 // and handles player speed and movement
 void physics(Game * game)
 {
-	
+
 	game->inAir(); 
-	game->applyGravity();
 	game->checkBottomScreen();
 
-	if(keys[XK_Left]) // left
+	if(!pausegame)
 	{
-		game->player.right = false;
-		game->player.left = true;
-		if(frames == 1)
-			frame += 0.125;
-		//cout << "left" << endl;
-		game->accelX(-1 * INITIAL_VELOCITY);
-	}
+		game->applyGravity();
 
-	if(keys[XK_Right]) // right
-	{
-		game->player.left = false;
-		game->player.right = true;
-		if(frames == 1)
-			frame +=  0.125;
-		//cout << "right" << endl;
- 		game->accelX(INITIAL_VELOCITY);
-	}
 
-	if(keys[XK_space] && game->if_jump) // spacebar
-	{
-		//cout << "jump" <<endl;
-		game->accelY(2 * INITIAL_VELOCITY);
-	}
-	 
-	if(killmovement) // kill movement on x axis only
-		game->player.velocity.x = 0;
+		if(keys[XK_Left]) // left
+		{
+			game->player.right = false;
+			game->player.left = true;
+			if(frames == 1)
+				frame += 0.125;
+			//cout << "left" << endl;
+			game->accelX(-1 * INITIAL_VELOCITY);
+		}
+
+		if(keys[XK_Right]) // right
+		{
+			game->player.left = false;
+			game->player.right = true;
+			if(frames == 1)
+				frame +=  0.125;
+			//cout << "right" << endl;
+			game->accelX(INITIAL_VELOCITY);
+		}
+
+		if(keys[XK_space] && game->if_jump) // spacebar
+		{
+			//cout << "jump" <<endl;
+			game->accelY(2 * INITIAL_VELOCITY);
+		}
+
+		if(killmovement) // kill movement on x axis only
+			game->player.velocity.x = 0;
+
 	
-	
-	
+
+
 	if(game->velX() > MAX_VELOCITY)
 		game->player.velocity.x = MAX_VELOCITY;
 	if(game->velX() < -1 * MAX_VELOCITY)
 		game->player.velocity.x = -1 * MAX_VELOCITY;
-		
-	
 
-	
-	game->move();
-	
-	
-	
+
+
+	if(!pausegame)
+	{
+		game->move();
+	}
+
+	}
 	int x_bubbler = 100;
 	int y_bubbler = window_height;
 
@@ -294,24 +316,24 @@ void physics(Game * game)
 			makeParticle(x_bubbler, y_bubbler);
 		}
 	}
-	
+
 	// particles
 	Particle *p = &par[numParticles];
-	
-	
-	
+
+
+
 	for(int i = 0; i < numParticles; ++i)
 	{
 		p = &par[i];
 		p->s.center.x += p->velocity.x;
 		p->s.center.y += p->velocity.y;
 		p->velocity.y -= 0.1; 
-	
+
 		if (p->s.center.y < 0.0 || p->s.center.y > window_height) 
 		{
 			//std::cout << "off screen" << std::endl;
 			memcpy(&par[i], &par[numParticles -1], 
-				sizeof(Particle));
+					sizeof(Particle));
 			numParticles--;
 		}
 	}
@@ -326,15 +348,15 @@ void render(Game * game)
 	glClearColor(0, 0, 0, 1.0);
 	glClear(GL_COLOR_BUFFER_BIT);
 
-	
+
 	// makes sure not to draw past window edges!
 	game->checkRightScreenHit();
 	game->checkLeftScreenHit();
-	
-	
-		float w, h;
+
+
+	float w, h;
 	int x,y;
-	
+
 	// texture
 	if(setbackground)
 	{
@@ -342,23 +364,23 @@ void render(Game * game)
 		glBindTexture(GL_TEXTURE_2D, backgroundTexture);
 		glBegin(GL_QUADS);
 		glTexCoord2f(0,1);
-			glVertex2i(0,0);
+		glVertex2i(0,0);
 		glTexCoord2f(0,0);
-			glVertex2i(0, window_height);
+		glVertex2i(0, window_height);
 		glTexCoord2f(1,0);
-			glVertex2i(window_width, window_height);
+		glVertex2i(window_width, window_height);
 		glTexCoord2f(1,1);
-			glVertex2i(window_width, 0);
+		glVertex2i(window_width, 0);
 		glEnd();
-		
-		
+
+
 		drawSpike();
-		
+
 		drawSkeleton(game);
-		
+
 	}
-	
-	
+
+
 	// TEXT
 	Rect r;
 	r.bot = window_height - 20;
@@ -368,14 +390,14 @@ void render(Game * game)
 	ggprint8b(&r, 16, 0x00FFFF00, "PhysicsRate: %i", static_cast<int>(1/physicsRate));
 	ggprint8b(&r, 16, 0x00FFFF00, "water particles: %i", numParticles);
 	ggprint8b(&r, 16, 0x00FFFF00, "Hit sides: %i", game->checkLeftScreenHit() || game->checkRightScreenHit());
-	
 
-	
+
+
 	if(!setbackground)
 	{
-		
+
 		//draw guy/rectangle
-	
+
 		glColor3ub(222,10,90);
 		glPushMatrix();
 		glTranslatef(game->player.position.x, game->player.position.y, 0);
@@ -388,7 +410,7 @@ void render(Game * game)
 		glVertex2i( w,-h);
 		glEnd();
 		glPopMatrix();
-		
+
 		//draw guy's (x,y) center coordinates
 		glColor3ub(0,0,255);
 		glPushMatrix();
@@ -426,8 +448,8 @@ void render(Game * game)
 		randColorWater+= 10; 
 	}
 
-	
 
-	
-	
+
+
+
 }
